requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
asyncpg==0.30.0
python-dotenv==1.2.2
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.11.1
